requirements.txt
aniso8601==8.0.0
astroid==2.4.2
autopep8==1.5.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
flask==1.1.2
flask-mongoengine==1.0.0
flask-pymongo==2.3.0
flask-restful==0.3.8
flask-wtf==0.14.3
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
jinja2==2.11.2
lazy-object-proxy==1.4.3
markupsafe==1.1.1
mccabe==0.6.1
mongoengine==0.21.0
pika==1.1.0
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pymongo==3.11.1
pytz==2020.4
six==1.15.0
toml==0.10.2
werkzeug==1.0.1
wrapt==1.12.1
wtforms==2.3.3
python-json-logger==2.0.1
PyYAML~=5.3.1